Update on "Teach Python TS frontend to parse complex literals"
Follow-up work:
1. Add complex support for the remaining binary ops as well as unary ops, and other registered ops
2. Extensive TS support for complex lists and dicts
3. Test all emitXYZ for all XYZ in `ir_emitter.cpp` (currently only emitConst, emitValueToTensor are tested). e.g., test loops etc.

[ghstack-poisoned]

+++ torch/csrc/jit/frontend/ir_emitter.cpp
@@ -667,7 +667,7 @@ struct to_ir {
   Function& method;
   std::shared_ptr<Graph> graph;
   ResolverPtr resolver;
-  std::unordered_map<int64_t, Value*, std::hash<int>> integral_constants;
+  std::unordered_map<int64_t, Value*, std::hash<int64_t>> integral_constants;
   std::unordered_map<double, Value*, std::hash<float>> fp_constants;
   std::unordered_map<
       c10::complex<double>,
